pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "peblo-quiz-backend"
version = "0.1.0"
description = "Adaptive quiz generation backend for the Peblo backend engineer challenge."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0",
  "asyncpg>=0.30.0",
  "fastapi>=0.116.0",
  "httpx>=0.28.0",
  "pydantic-settings>=2.7.0",
  "pypdf>=5.3.0",
  "python-dotenv>=1.0.1",
  "python-multipart>=0.0.20",
  "sqlalchemy>=2.0.38",
  "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
  "aiosqlite>=0.21.0",
  "pytest>=8.3.4",
  "pytest-asyncio>=0.25.3",
  "pytest-cov>=6.0.0",
  "ruff>=0.11.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["alembic", "app/models", "app/core/database.py"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["UP037", "UP042"]

[tool.ruff.lint.per-file-ignores]
"app/http/routes.py" = ["B008", "E501"]
"app/db/models.py" = ["E501", "I001"]
"app/db/session.py" = ["E501", "I001"]
"app/logic/ai.py" = ["E501"]
"app/logic/scoring.py" = ["E501"]
"app/services/chunking.py" = ["I001"]
"tests/test_quiz_flow.py" = ["E501"]
